Added a base class for Dipole BSSRDFs and refactored common code.
