+++ src/appleseed/renderer/modeling/bssrdf/dipolebssrdf.cpp
@@ -1,0 +1,215 @@
+
+//
+// This source file is part of appleseed.
+// Visit http://appleseedhq.net/ for additional information and resources.
+//
+// This software is released under the MIT license.
+//
+// Copyright (c) 2015 Francois Beaune, The appleseedhq Organization
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+//
+
+// Interface header.
+#include "dipolebssrdf.h"
+
+// appleseed.renderer headers.
+#include "renderer/modeling/bssrdf/sss.h"
+#include "renderer/modeling/input/inputevaluator.h"
+
+// appleseed.foundation headers.
+#include "foundation/math/scalar.h"
+#include "foundation/math/vector.h"
+#include "foundation/utility/containers/dictionary.h"
+#include "foundation/utility/containers/specializedarrays.h"
+#include "foundation/utility/memory.h"
+
+// Standard headers.
+#include <cmath>
+#include <cstddef>
+
+using namespace foundation;
+using namespace std;
+
+namespace renderer
+{
+
+//
+// Dipole BSSRDF base class implementation.
+//
+
+DipoleBSSRDF::DipoleBSSRDF(
+    const char*             name,
+    const ParamArray&       params)
+  : BSSRDF(name, params)
+{
+    m_inputs.declare("weight", InputFormatScalar, "1.0");
+    m_inputs.declare("reflectance", InputFormatSpectralReflectance);
+    m_inputs.declare("reflectance_multiplier", InputFormatScalar, "1.0");
+    m_inputs.declare("dmfp", InputFormatScalar, "5.0");
+    m_inputs.declare("dmfp_multiplier", InputFormatScalar, "0.1");
+    m_inputs.declare("anisotropy", InputFormatScalar);
+    m_inputs.declare("outside_ior", InputFormatScalar);
+    m_inputs.declare("inside_ior", InputFormatScalar);
+}
+
+size_t DipoleBSSRDF::compute_input_data_size(
+    const Assembly&         assembly) const
+{
+    return align(sizeof(DipoleBSSRDFInputValues), 16);
+}
+
+void DipoleBSSRDF::evaluate_inputs(
+    const ShadingContext&   shading_context,
+    InputEvaluator&         input_evaluator,
+    const ShadingPoint&     shading_point,
+    const size_t            offset) const
+{
+    BSSRDF::evaluate_inputs(shading_context, input_evaluator, shading_point, offset);
+
+    DipoleBSSRDFInputValues* values =
+        reinterpret_cast<DipoleBSSRDFInputValues*>(input_evaluator.data() + offset);
+
+    // Apply multipliers.
+    values->m_reflectance *= static_cast<float>(values->m_reflectance_multiplier);
+    values->m_dmfp *= values->m_dmfp_multiplier;
+
+    // Clamp reflectance to [0, 1].
+    values->m_reflectance = saturate(values->m_reflectance);
+
+#if 1
+    // Compute sigma_a and sigma_s from the reflectance and dmfp parameters.
+    compute_absorption_and_scattering(
+        values->m_reflectance,
+        values->m_dmfp,
+        values->m_inside_ior / values->m_outside_ior,
+        values->m_anisotropy,
+        values->m_sigma_a,
+        values->m_sigma_s,
+        values->m_sigma_tr);
+#else
+    // Skim milk.
+    values->m_sigma_a = Color3f(0.0014f, 0.0025f, 0.0142f) * 1000.0f;
+    values->m_sigma_s = Color3f(0.70f, 1.22f, 1.90f) * 1000.0f;
+    values->m_anisotropy = 0.0;
+
+    effective_extinction_coefficient(
+        values->m_sigma_a,
+        values->m_sigma_s,
+        values->m_anisotropy,
+        values->m_sigma_tr);
+#endif
+}
+
+
+//
+// DipoleBSSRDFFactory class implementation.
+//
+
+DictionaryArray DipoleBSSRDFFactory::get_input_metadata() const
+{
+    DictionaryArray metadata;
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "weight")
+            .insert("label", "Weight")
+            .insert("type", "colormap")
+            .insert("entity_types",
+                Dictionary().insert("texture_instance", "Textures"))
+            .insert("use", "optional")
+            .insert("default", "1.0"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "reflectance")
+            .insert("label", "Diffuse Surface Reflectance")
+            .insert("type", "colormap")
+            .insert("entity_types",
+                Dictionary()
+                    .insert("color", "Colors")
+                    .insert("texture_instance", "Textures"))
+            .insert("use", "required")
+            .insert("default", "0.5"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "reflectance_multiplier")
+            .insert("label", "Diffuse Surface Reflectance Multiplier")
+            .insert("type", "colormap")
+            .insert("entity_types",
+                Dictionary().insert("texture_instance", "Textures"))
+            .insert("use", "optional")
+            .insert("default", "1.0"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "dmfp")
+            .insert("label", "Diffuse Mean Free Path")
+            .insert("type", "colormap")
+            .insert("entity_types",
+                Dictionary()
+                    .insert("texture_instance", "Textures"))
+            .insert("use", "required")
+            .insert("default", "5"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "dmfp_multiplier")
+            .insert("label", "Diffuse Mean Free Path Multiplier")
+            .insert("type", "colormap")
+            .insert("entity_types",
+                Dictionary().insert("texture_instance", "Textures"))
+            .insert("use", "optional")
+            .insert("default", "0.1"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "anisotropy")
+            .insert("label", "Anisotropy")
+            .insert("type", "numeric")
+            .insert("min_value", "-1.0")
+            .insert("max_value", "1.0")
+            .insert("use", "required")
+            .insert("default", "0.0"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "outside_ior")
+            .insert("label", "Outside Index of Refraction")
+            .insert("type", "numeric")
+            .insert("min_value", "0.0")
+            .insert("max_value", "5.0")
+            .insert("use", "required")
+            .insert("default", "1.0"));
+
+    metadata.push_back(
+        Dictionary()
+            .insert("name", "inside_ior")
+            .insert("label", "Inside Index of Refraction")
+            .insert("type", "numeric")
+            .insert("min_value", "0.0")
+            .insert("max_value", "5.0")
+            .insert("use", "required")
+            .insert("default", "1.3"));
+
+    return metadata;
+}
+
+}   // namespace renderer
